[DC] I've pulled the implementation of the declarations and fractions controller methods up into the respective traits. The traits define abstract methods to get their connectors. I've moved the objects that make up the live implementations into a `controllers.live` sub-package and overridden the controller methods to return `NotImplemented`.
The sandboxes look for stub connector definitions in the conf file, whilst the live controllers currently don't provide connectors as they should always return `NotImplemented` in any case.

All of this is to unify the way the controllers work so that we can build the implementations through experimentation with the sandboxes and hopefully ensure they'll behave identically with live backend services.